+++ cdc/sinkv2/ddlsink/mq/ddlproducer/kafka_ddl_producer_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package producer
+package ddlproducer
 
 import (
 	"context"
